requirements.txt
asgiref==3.7.2
Django==5.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-rest-passwordreset==1.3.0
django-user-agents==0.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
mysqlclient==2.2.0
packaging==23.2
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
ua-parser==0.18.0
uritemplate==4.1.1
user-agents==2.2.0
